chore: update mutation paths
