+++ packages/query/src/index.ts
@@ -13,8 +13,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-export { useMutationWallet } from "./useMutationWallet";
-export { useMutationWalletSettings } from "./useMutationWalletSettings";
+export { useMutationWalletUpdate } from "./useMutationWalletUpdate";
+export { useMutationWalletSettingsUpdate } from "./useMutationWalletSettingsUpdate";
 export { useQueryActivities } from "./useQueryActivities";
 export { useQueryActivitiesCount } from "./useQueryActivitiesCount";
 export { useQueryAuthSession } from "./useQueryAuthSession";
