+++ packages/query/src/useMutationWalletSettingsUpdate.ts
@@ -13,11 +13,10 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { updateWallet } from "@lightdotso/client";
-import type { WalletData } from "@lightdotso/data";
+import { updateWalletSettings } from "@lightdotso/client";
+import type { WalletSettingsData } from "@lightdotso/data";
 import type { WalletParams } from "@lightdotso/params";
 import { queryKeys } from "@lightdotso/query-keys";
-import { useAuth } from "@lightdotso/stores";
 import { toast } from "@lightdotso/ui";
 import { useQueryClient, useMutation } from "@tanstack/react-query";
 
@@ -25,13 +24,7 @@ import { useQueryClient, useMutation } from "@tanstack/react-query";
 // Query Mutation
 // -----------------------------------------------------------------------------
 
-export const useMutationWallet = (params: WalletParams) => {
-  // ---------------------------------------------------------------------------
-  // Stores
-  // ---------------------------------------------------------------------------
-
-  const { clientType } = useAuth();
-
+export const useMutationWalletSettingsUpdate = (params: WalletParams) => {
   // ---------------------------------------------------------------------------
   // Query
   // ---------------------------------------------------------------------------
@@ -39,59 +32,58 @@ export const useMutationWallet = (params: WalletParams) => {
   const queryClient = useQueryClient();
 
   const { mutate, isPending, isSuccess, isError } = useMutation({
-    mutationFn: async (data: Partial<WalletData>) => {
-      const loadingToast = toast.loading("Updating name...");
+    mutationFn: async (data: WalletSettingsData) => {
+      const loadingToast = toast.loading("Updating wallet settings...");
 
-      const res = await updateWallet(
-        {
-          params: {
-            query: {
-              address: params.address,
-            },
-          },
-          body: {
-            name: data.name,
+      const res = await updateWalletSettings({
+        params: {
+          query: {
+            address: params.address,
           },
         },
-        clientType,
-      );
+        body: {
+          wallet_settings: {
+            is_enabled_testnet: data.is_enabled_testnet,
+          },
+        },
+      });
 
       toast.dismiss(loadingToast);
 
       // Return if the response is 200
       res.match(
         _ => {
-          toast.success("Successfully updated name.");
+          toast.success("Successfully updated wallet settings.");
         },
         err => {
           if (err instanceof Error) {
             toast.error(err.message);
           } else {
-            toast.error("Failed to update name.");
+            toast.error("Something went wrong.");
           }
-
-          throw err;
         },
       );
     },
     // When mutate is called:
-    onMutate: async (wallet: Partial<WalletData>) => {
+    onMutate: async (walletSettings: WalletSettingsData) => {
       // Cancel any outgoing refetches
       // (so they don't overwrite our optimistic update)
       await queryClient.cancelQueries({
-        queryKey: queryKeys.wallet.get({ address: params.address }).queryKey,
+        queryKey: queryKeys.wallet.settings({ address: params.address })
+          .queryKey,
       });
 
       // Snapshot the previous value
-      const previousSettings: WalletData | undefined = queryClient.getQueryData(
-        queryKeys.wallet.get({ address: params.address }).queryKey,
-      );
+      const previousSettings: WalletSettingsData | undefined =
+        queryClient.getQueryData(
+          queryKeys.wallet.settings({ address: params.address }).queryKey,
+        );
 
       // Optimistically update to the new value
       queryClient.setQueryData(
         queryKeys.wallet.settings({ address: params.address }).queryKey,
-        (old: WalletData) => {
-          return { ...old, wallet };
+        (old: WalletSettingsData) => {
+          return { ...old, walletSettings };
         },
       );
 
@@ -101,32 +93,30 @@ export const useMutationWallet = (params: WalletParams) => {
     // If the mutation fails, use the context we returned above
     onError: (err, _newWalletSettings, context) => {
       queryClient.setQueryData(
-        queryKeys.wallet.get({ address: params.address }).queryKey,
+        queryKeys.wallet.settings({ address: params.address }).queryKey,
         context?.previousSettings,
       );
 
       if (err instanceof Error) {
         toast.error(err.message);
       } else {
-        toast.error("Failed to update name.");
+        toast.error("Something went wrong.");
       }
     },
     onSettled: () => {
       queryClient.invalidateQueries({
-        queryKey: queryKeys.wallet.get({ address: params.address }).queryKey,
+        queryKey: queryKeys.wallet.settings({ address: params.address })
+          .queryKey,
       });
       queryClient.invalidateQueries({
-        queryKey: queryKeys.wallet.list({
-          address: params.address,
-          limit: Number.MAX_SAFE_INTEGER,
-          offset: 0,
-        }).queryKey,
+        queryKey: queryKeys.auth.session({ address: params.address }).queryKey,
       });
 
       // Invalidate the cache for the address
-      // fetch(`/api/revalidate/tag?tag=${address}`);
+      fetch(`/api/revalidate/tag?tag=${params.address}`);
     },
-    mutationKey: queryKeys.wallet.get({ address: params.address }).queryKey,
+    mutationKey: queryKeys.wallet.settings({ address: params.address })
+      .queryKey,
   });
 
   return {
